pyproject.toml
[project]
name = "rocketwatch"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "aiohttp==3.13.3",
    "aiocache==0.12.3",
    "anthropic==0.84.0",
    "anyascii==0.3.3",
    "beautifulsoup4==4.14.3",
    "bidict==0.23.1",
    "cachetools==7.0.3",
    "colorama==0.4.6",
    "cronitor==4.9.0",
    "dice==4.0.0",
    "discord.py==2.7.1",
    "eth-typing==5.2.1",
    "eth-utils==5.3.1",
    "etherscan_labels @ git+https://github.com/haloooloolo/etherscan-labels",
    "graphql_query==1.4.0",
    "hexbytes==1.3.1",
    "humanize==4.15.0",
    "inflect==7.5.0",
    "matplotlib==3.10.8",
    "numpy==2.4.2",
    "pillow==12.1.1",
    "psutil==7.2.2",
    "pydantic>=2.0.0,<3.0.0",
    "pymongo==4.16.0",
    "python_i18n==0.3.9",
    "pytz==2026.1.post1",
    "regex==2026.2.28",
    "retry-async==0.1.4",
    "seaborn==0.13.2",
    "tabulate==0.10.0",
    "termplotlib==0.3.9",
    "tiktoken==0.12.0",
    "uptime==3.0.1",
    "web3>=7.0.0,<8.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=1.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = ["B", "E", "F", "I", "RUF", "SIM", "UP", "W"]
ignore = ["E501", "E203", "E231"]

[tool.ruff.lint.isort]
known-first-party = ["utils", "plugins", "strings", "rocketwatch"]
